requirements.txt
django == 4.1.2
elastic-apm == 6.12.0
psutil == 5.9.2 
django-mailer == 2.2
djangorestframework == 3.14.0
django-cors-headers == 3.13.0
drf_yasg == 1.21.4
PyYAML == 6.0
django-cron == 0.6.0